requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
itsdangerous>=2.1.0
httpx>=0.26.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
